move SnapshotConfig into its own module; add snapshot_runtime_info to SnapshotConfig; move ArchiveFormat to snapshot_utils
